intern/maze: use 'double' to represent heuristic more precisely
In a very few cases this results in longer paths and in few other cases
this results in increased expansion of nodes but in most cases this
resulted in the shortest length path with fewer node expansion.

+++ intern/files/maze/a_star/frontier/pqueue.c
@@ -42,7 +42,7 @@ inline static void swap_nodes(struct heap_elem **m, struct heap_elem **n)
  *
  * Returns 0 on success and non-zero value on failure.
  */
-static int heap_decrease_key(struct min_heap *mheap, size_t elem_offset, unsigned new_key)
+static int heap_decrease_key(struct min_heap *mheap, size_t elem_offset, double new_key)
 {
 #ifdef KS_PRIORITY_QUEUE_DEBUG
 	if (mheap == NULL)
